Reintroduce DataPreviewer and start modifying visualisation logic to support dataset and distribution previews
